requirements.txt
appdirs==1.4.0
appnope==0.1.0
boto==2.46.1
boto3==1.4.4
botocore==1.5.19
decorator==4.0.11
defusedxml==0.5.0rc1
Django==1.10.5
django-braces==1.11.0
django-filter==1.0.1
django-oauth-toolkit==0.11.0
django-rest-framework-social-oauth2==1.0.5
django-ses==0.8.1
django-smtp-ssl==1.0
djangorestframework==3.5.3
docutils==0.13.1
ipython==5.2.2
ipython-genutils==0.1.0
jmespath==0.9.1
Markdown==2.6.8
oauthlib==1.1.2
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
PyJWT==1.4.2
pyparsing==2.1.10
python-dateutil==2.6.0
python3-openid==3.0.10
requests==2.13.0
requests-oauthlib==0.7.0
s3transfer==0.1.10
simplegeneric==0.8.1
six==1.10.0
social-auth-app-django==1.0.1
social-auth-core==1.1.0
traitlets==4.3.1
wcwidth==0.1.7
python-igraph==0.7.1
